fix flaky tests on main
