+++ package.json
@@ -25,7 +25,7 @@
     "@babel/parser": "^7.28.0",
     "@babel/traverse": "^7.28.0",
     "@eslint/js": "^8.57.0",
-    "@playwright/test": "^1.52.0",
+    "@playwright/test": "^1.55.0",
     "allure-commandline": "^2.34.0",
     "allure-playwright": "^3.2.1",
     "eslint": "^9.18.0",
